+++ datafusion-flight-sql-server/examples/test.csv
@@ -1,0 +1,4 @@
+foo,bar
+a,1
+b,2
+c,3
