--- datafusion-flight-sql-server/src/server/mod.rs
@@ -1,6 +1,0 @@
-mod service;
-pub use service::*;
-mod state;
-pub use state::*;
-mod session;
-pub use session::*;
