+++ datafusion-flight-sql-table-provider/Cargo.toml
@@ -1,16 +1,13 @@
 [package]
-name = "datafusion-flight-sql-server"
+name = "datafusion-flight-sql-table-provider"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
-
-[lib]
-name = "datafusion_flight_sql_server"
-path = "src/lib.rs"
+readme.workspace = true
 
 [dependencies]
+async-trait.workspace = true
 datafusion.workspace = true
-datafusion-substrait.workspace = true
 # XXX use the release verion on crates.io 
 datafusion-federation = { path = "../datafusion-federation", features = [
     "sql",
@@ -23,11 +20,5 @@ tonic = { version = "0.11.0", features = [
     "codegen",
     "prost",
 ] }
-prost = "0.12.3"
 arrow = "52.0.0"
 arrow-flight = { version = "52.2.0", features = ["flight-sql-experimental"] }
-log = "0.4.22"
-
-[dev-dependencies]
-tokio = { version = "1.39.3", features = ["full"] }
-datafusion-flight-sql-table-provider = { path = "../datafusion-flight-sql-table-provider" }
